+++ libs/core/cfabric/features/computed.py
@@ -1,9 +1,9 @@
 """
 # Pre-computed data components
 
-In order to make the API work, TF prepares some data and saves it in
+In order to make the API work, CF prepares some data and saves it in
 quick-load format. Most of this data are the features, but there is some extra
-data needed for the special functions of the `tf.parameters.WARP` features and the
+data needed for the special functions of the `cfabric.parameters.WARP` features and the
 `cfabric.locality.Locality` API.
 
 Normally, you do not use this data, but since it is there, it might be valuable,
@@ -18,8 +18,8 @@ so we have made it accessible in the `cfabric.computed.Computeds`-API.
     - Reduced memory footprint
     - Near-zero startup time after initial compilation
 
-    Use `TF.compile()` to generate the `.cfm` directory from `.tf` source files.
-    Subsequent calls to `TF.load()` will automatically use the compiled format.
+    Use `CF.compile()` to generate the `.cfm` directory from `.tf` source files.
+    Subsequent calls to `CF.load()` will automatically use the compiled format.
 """
 
 from __future__ import annotations